pyproject.toml
[project]
name = "k8s-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Zakaria Jaadi",email = "mr.zakaria.jaadi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "mlflow (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "hyperopt (>=0.2.7,<0.3.0)",
    "deepchecks (>=0.19.1,<0.20.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "prefect[kubernetes] (>=3.4.7,<4.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "numpy (<2.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "scipy==1.15.3",
    "statsmodels==0.14.4",
    "boto3 (>=1.39.4,<2.0.0)",
    "graphql-core (==3.1.7)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
